requirements.txt
# Core Slack SDK and web framework
slack-sdk==3.26.0
slack-bolt-python==1.18.0
flask==3.0.0
gunicorn==21.2.0

# HTTP client and API interactions
requests==2.31.0
aiohttp==3.9.0

# Database drivers
psycopg2-binary==2.9.9
pymongo==4.6.0
redis==5.0.1

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0

# Date/time handling
python-dateutil==2.8.2
pytz==2023.3

# Logging and monitoring
structlog==23.2.0
sentry-sdk==1.38.0

# Data validation and parsing
marshmallow==3.20.0
jsonschema==4.20.0

# Utilities
uuid==1.30
click==8.1.7

# Development dependencies
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.0

# Optional: Machine learning and NLP utilities
numpy==1.25.0
pandas==2.1.0 